@@ -102,28 +102,35 @@
     "karma-webpack": "3.0.0",
     "less": "3.0.4",
     "less-loader": "4.1.0",
-    "lint-staged": "7.1.2",
-    "mini-css-extract-plugin": "0.4.0",
+    "lint-staged": "7.2.0",
+    "mini-css-extract-plugin": "0.4.1",
     "mocha": "5.2.0",
-    "pofile": "1.0.10",
-    "prettier": "1.12.1",
+    "pofile": "1.0.11",
+    "prettier": "1.13.7",
     "raw-loader": "0.5.1",
-    "react-test-renderer": "16.4.0",
+    "react-test-renderer": "16.4.1",
     "remap-istanbul": "0.11.1",
     "style-loader": "0.21.0",
-    "ts-loader": "4.3.0",
+    "ts-loader": "4.4.2",
     "tslint": "5.10.0",
-    "typescript": "2.8.3",
+    "tslint-sonarts": "1.7.0",
+    "typescript": "2.9.2",
     "url-loader": "1.0.1",
-    "webpack": "4.8.3",
+    "webpack": "4.15.0",
     "webpack-bundle-analyzer": "2.13.1",
-    "webpack-cli": "2.1.4",
+    "webpack-cli": "3.0.8",
     "webpack-dev-server": "3.1.4"
   },
   "lint-staged": {
-    "ignore": ["**/*.min.js", "**/*.min.css"],
+    "ignore": [
+      "**/*.min.js",
+      "**/*.min.css"
+    ],
     "linters": {
-      "*.{ts,tsx,js,json,css,less}": ["prettier --write", "git add"]
+      "*.{ts,tsx,js,json,css,less}": [
+        "prettier --write",
+        "git add"
+      ]
     }
   }
 }
